Solution#3

+++ app/main.py
@@ -19,18 +19,19 @@ def move_file(command: str) -> None:
             os.remove(file_cp)
             return
         if len(file_path.split("/")) == 1:
-            root, ext_f = os.path.splitext(file_path)
-            root1, ext_cp = os.path.splitext(file_path)
-            if ext_f == ext_cp:
+            if os.path.exists(file_cp):
                 os.rename(file_cp, file_path)
             return
         file_path_split = file_path.split("/")
-        cp_file_txt = file_path_split.pop(-1)
+        file_ex = os.path.basename(file_path)
+        file_path_split.remove(file_ex)
         file_path = "/".join(file_path_split)
+        if not os.path.exists(file_cp):
+            return
         if not os.path.exists(file_path):
             os.makedirs(file_path)
         with (open(file_cp, "r") as f,
-              open(f"{file_path}/{cp_file_txt}", "w") as f_cp):
+              open(f"{file_path}/{file_ex}", "w") as f_cp):
             for line in f:
                 f_cp.write(line)
         os.remove(file_cp)
